requirements.txt
asgiref==3.7.2
Django==5.0.2
django-crispy-forms==2.1
django-filter==23.5
djangorestframework==3.14.0
nullable==0.2.0
pillow==10.2.0
psycopg==3.1.18
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
djangorestframework-simplejwt==5.3.1
